+++ src/library/DIPS.Mobile.UI/API/Camera/ImageCapturing/Settings/ImageCaptureSettings.cs
@@ -1,4 +1,5 @@
 using System.Windows.Input;
+using DIPS.Mobile.UI.Resources.LocalizedStrings.LocalizedStrings;
 
 namespace DIPS.Mobile.UI.API.Camera.ImageCapturing.Settings;
 
@@ -9,6 +10,13 @@ public class ImageCaptureSettings
     /// </summary>
     /// <remarks>The done button will be hidden if this is not set</remarks>
     public ICommand? DoneButtonCommand { get; set; }
+    
+    /// <summary>
+    /// The text to be displayed on the done button
+    /// </summary>
+    /// <remarks>Default is <see cref="DUILocalizedStrings.Cancel"/></remarks>
+    public string DoneButtonText { get; set; } = DUILocalizedStrings.Cancel;
+    
     public PostCaptureAction PostCaptureAction { get; set; }
     public int? MaxHeightOrWidth { get; set; }
     public bool CanChangeMaxHeightOrWidth { get; set; }