+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [37.0.1]
+- [ImageCapture] Changed text from "Take new photo" to "Retake photo".
+- [ImageCapture] Changed text from "Done" to "Cancel".
+- [ImageCapture] Changed icon from "information_line" to "information_fill".
+- [ImageCapture] Changed text from "Remove" to "Delete".
+
 ## [37.0.0]
 - [ImageCapture] Added a new `ImageCapture` API along side with `Camera` api that can be used for both image capture and bar code scanning.
 - [BarcodeScanner][BreakingChange] Requires you to provide a CameraFailed delegate when starting the barcode scanner.